Split ieee to math_functions, next, parse and precision
